% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kyphosis.R
\docType{data}
\name{kyphosis}
\alias{kyphosis}
\title{Data on Children who have had Corrective Spinal Surgery}
\format{
This data frame contains the following columns:
\describe{
  \item{\code{Kyphosis}}{
    a factor with levels
    \code{absent} 
    \code{present}
    indicating if a kyphosis (a type of deformation)
    was present after the operation.
  }
  \item{\code{Age}}{
    in months
  }
  \item{\code{Number}}{
    the number of vertebrae involved 
  }
  \item{\code{Start}}{
    the number of the first (topmost) vertebra operated on.
  }
}
}
\source{
Chambers and Hastie (1992).
}
\usage{
data(kyphosis)
}
\description{
The \code{kyphosis} data frame has 81 rows and four columns representing data
on children who have had corrective spinal surgery.
}
\details{
Provided with the original package
}
\examples{
## Load the data set:
data("kyphosis")

## Number of rows and columns in the data table:
dim(kyphosis)

## First six rows of the data table:
head(kyphosis)

## Summary of the data table:
summary(kyphosis)

## A few plots of the data within the data frame:
plot(Number ~ Age, data=kyphosis, las=1L)
plot(Start ~ as.factor(Number), data=kyphosis, las=1L)
plot(Age/12 ~ Kyphosis, data=kyphosis, las=1L, ylab="Age (years)")
plot(Number ~ Kyphosis, data=kyphosis, las=1L)
plot(Age/12 ~ Start, data=kyphosis, las=1L, ylab="Age (years)")

}
\references{
John M. Chambers and Trevor J. Hastie eds. (1992) Statistical Models in S,
 Wadsworth and Brooks/Cole, Pacific Grove, CA 1992.
}
\keyword{datasets}
