% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvpart-package.R
\name{mvpart-package}
\alias{mvpart-package}
\title{\packageTitle{mvpart}}
\description{
\packageDescription{mvpart}


The DESCRIPTION file:
\packageDESCRIPTION{mvpart}
\packageIndices{mvpart}
}
\author{
\packageAuthor{mvpart}
Maintainer: \packageMaintainer{mvpart}
}
