% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/printcp.R
\name{printcp}
\alias{printcp}
\title{Displays CP table for an Rpartmv Object}
\usage{
printcp(x, digits = getOption("digits") - 2L)
}
\arguments{
\item{x}{An \code{\link{rpartmv-class}} object.}

\item{digits}{The number of digits of numbers to print.}
}
\value{
The \code{\link{rpartmv-class}} object's cp table.
}
\description{
Displays the \code{cp} table for fitted \code{rpartmv} object.
}
\details{
Prints a table of optimal pruning based on a complexity parameter.
}
\examples{
## Load the car data set:
data("car.test.frame")

## Estimating a regression tree with a single descriptor:
rt.auto1 <- rpartmv(Mileage ~ Weight, data=car.test.frame)

## Print the model:
rt.auto1

## Print the cp table:
printcp(rt.auto1)

}
\seealso{
\code{\link{summary.rpartmv}}
}
\keyword{tree}
