% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{rpartmv-generics}
\alias{rpartmv-generics}
\alias{prune}
\alias{rsq}
\alias{meanvar}
\title{Generic Methods}
\usage{
prune(object, ...)

rsq(object, ...)

meanvar(object, ...)
}
\arguments{
\item{object}{An object of a class for which the method is implemented
(e.g., \code{rpartmv-class} object).}

\item{...}{Optional arguments to be passed internally to other functions.}
}
\value{
The return value depends on the implementation of the methods for a
particular S3 class.
}
\description{
A set of non-standard generic functions used by package mvpart.
}
\section{Functions}{
\itemize{
\item \code{prune()}: Prune Method

Method to prune branch for a regression or classification tree

\item \code{rsq()}: R-square Method

Method to generate the R-square variation plot of an object.

\item \code{meanvar()}: Meanvar method

Method to generate the mean-variance plot for an object.

}}
\seealso{
\code{prune.rpartmv} \code{rsq.rpartmv}, and \code{meanvar.rpartmv}
}
