% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rpartmv.pca.R
\name{rpartmv.pca}
\alias{rpartmv.pca}
\title{Principal Components Plot of a Multivariate Rpartmv Object}
\usage{
rpartmv.pca(
  tree,
  pts = TRUE,
  plt.allx = TRUE,
  speclabs = TRUE,
  specvecs = TRUE,
  wgt.ave = FALSE,
  add.tree = TRUE,
  cv1 = 1L,
  cv2 = 2L,
  chulls = TRUE,
  interact = FALSE,
  ...
)
}
\arguments{
\item{tree}{A fitted object of class \code{rpartmv} containing a multivariate
regression tree.}

\item{pts}{If \code{TRUE}, large points representing the leaf means are
plotted.}

\item{plt.allx}{If \code{TRUE}, small points representing individual cases
are plotted.}

\item{speclabs}{If \code{TRUE} the labels of the response variables are
plotted.}

\item{specvecs}{If \code{TRUE} the vectors of the response variables are
plotted provided \code{wgt.ave} is \code{FALSE}.}

\item{wgt.ave}{If \code{TRUE} use weighted averages of responses not vectors.}

\item{add.tree}{If \code{TRUE} add the tree structure to the plot.}

\item{cv1}{Defines the principal component to plot horizontally -- but see
\code{interact}.}

\item{cv2}{Defines the principal component to plot vertically -- but see
\code{interact}.}

\item{chulls}{If \code{TRUE} adds convex hulls to thr tree groups.}

\item{interact}{If \code{TRUE} the plot can be viewed in dimensions by
left-clicking to top-left, bottom-right or bottom-left (reset).}

\item{...}{Arguments to be passed to or from other functions or methods.}
}
\value{
\code{NULL} (well, this is not really true).
}
\description{
Plots a PCA of the rpartmv object on the current graphics device.
}
\details{
This function plots a PCA biplot of the group means (leaves) of
multivariate regresssion objects of class \code{rpartmv}. The responses and
group means and indivdual cases can be shown on the plot. If responses are
positive (e.g., species-environment data) weighted averages of responses can
be plotted.

The PCA biplot plot is produced on the current graphics device.
}
\examples{
## Load the spider data set:
data(spider)

## Splitting the table into the responses (species, Y)
## and the descriptors (X):
Y <- data.matrix(spider[,1L:12L])
X <- spider[,13L:18L]
## Note: the multivariate response needs to procesed using function data.matrix
## before being used as a multivariate response by function rpartmv.

## Multivariate regression tree on the spider species response matrix and all
## the descriptors in the spider data set.

## Using the default settings:
spider1 <- mvpart(Y ~ herbs + reft + moss + sand + twigs + water, data=X)

## Non-interactive PCA plot:
rpartmv.pca(spider1)

## Interactive PCA plot:
rpartmv.pca(spider1, wgt.ave=TRUE, interact=TRUE)

}
\seealso{
\code{\link{rpartmv}}
}
\keyword{tree}
