% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rpconvert.R
\name{rpconvert}
\alias{rpconvert}
\title{Update an Rpartmv Object}
\usage{
rpconvert(x)
}
\arguments{
\item{x}{An \code{\link{rpartmv-class}} object.}
}
\value{
An updated rpartmv object
}
\description{
Objects of class rpartmv changed (slightly) in their internal format in order
to accommodate the changes for user-written split functions.  This routine
updates an old object to the new format.
}
\seealso{
\code{\link{rpartmv}}
}
\keyword{tree}
