% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snip.rpartmv.R
\name{snip.rpartmv}
\alias{snip.rpartmv}
\title{Snip Subtrees of an Rpartmv Object}
\usage{
snip.rpartmv(x, toss)
}
\arguments{
\item{x}{An \code{\link{rpartmv-class}} object.}

\item{toss}{An integer vector containing indices (node numbers) of all
sub trees to be snipped off. If missing, user selects branches to snip off as
described below.}
}
\value{
A \code{\link{rpartmv-class}} object containing the nodes that
remain after specified or selected sub trees have been snipped off.
}
\description{
Creates a "snipped" \code{\link{rpartmv-class}} object, containing the nodes
that remain after selected sub trees have been snipped off. The user can snip
nodes using the toss argument, or interactively by clicking the mouse button
on specified nodes within the graphics window.
}
\details{
A dendrogram of \code{rpartmv} is expected to be visible on the
graphics device, and a graphics input device (e.g., a mouse) is required.
Clicking (the selection button) on a node displays the node number, sample
size, response yvalue, and Error (dev). Clicking a second time on the same
node snips that sub tree off and visually erases the sub tree. This process
may be repeated an number of times. Warnings result from selecting the root
or leaf nodes.  Clicking the exit button will stop the snipping process and
return the resulting \code{\link{rpartmv-class}} object.

Warning: visually erasing the plot is done by over-plotting with the
background colour. This will do nothing if the background is transparent
(often true for screen devices).

See the documentation for the specific graphics device for details on
graphical input techniques.
}
\seealso{
\code{\link{plot.rpartmv}}
}
\keyword{tree}
