% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdist.R
\name{gdist}
\alias{gdist}
\title{Dissimilarity Measures}
\usage{
gdist(x, method = "bray", keepdiag = FALSE, full = FALSE, sq = FALSE)
}
\arguments{
\item{x}{Data matrix.}

\item{method}{Dissimilarity index. Choices are: "manhattan", "euclidean",
"canberra", "bray", "kulczynski", "gower", "maximum", "binary", "chisq",
"chord", "beta0", "beta1", or "beta2"}

\item{keepdiag}{Compute amd keep diagonals.}

\item{full}{Return the square dissimilarity matrix.}

\item{sq}{Square the dissimilarities -- useful for distance-based
partitioning.}
}
\value{
Should be interchangeable with \code{\link[stats]{dist}} and returns
a distance object of the same type.
}
\description{
The function computes useful dissimilarity indices which are known to have a
good rank-order relation with gradient separation and are thus efficient in
community ordination with multidimensional scaling.
}
\details{
Infamous ''double zeros'' are removed in Canberra dissimilarity.

Euclidean and Manhattan dissimilarities are not good in gradient separation
without proper standardization but are still included for comparison and
special needs.

Some of indices become identical or rank-order similar after some
standardization.
}
\note{
The function is an alternative to \code{\link[stats]{dist}} adding some
ecologically meaningful indices.  Both methods should produce similar types
of objects which can be interchanged in any method accepting either.
Manhattan and Euclidean dissimilarities should be identical in both methods,
and Canberra dissimilarity may be similar.
}
\references{
Faith, D.P, Minchin, P.R. and Belbin, L. (1987) Compositional dissimilarity
as a robust measure of ecological distance. \emph{Vegetatio} 69, 57-68.
}
\author{
Jari Oksanen  -- modified Glenn De'ath (Dec 03)
}
\keyword{multivariate}
