% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotcp.R
\name{plotcp}
\alias{plotcp}
\title{Plot a Complexity Parameter Table for an Rpartmv Fit}
\usage{
plotcp(
  x,
  xvse = 1,
  minline = TRUE,
  lty = 3,
  col = 1,
  upper = c("size", "splits", "none"),
  tab,
  resub.err = TRUE,
  adj.df = FALSE,
  ...
)
}
\arguments{
\item{x}{An object of class \code{rpartmv}.}

\item{xvse}{Multiplier for xvse * SE above the minimum of the curve.}

\item{minline}{Whether a horizontal line is drawn 1SE above the minimum of
the curve.}

\item{lty}{Type of lines used in plot, default = 3 (a dotted line).}

\item{col}{Color of lines, default = 1 (black).}

\item{upper}{What is plotted on the top axis: the size of the tree (the
number of leaves), the number of splits or nothing.}

\item{tab}{Used for multiple cross-validation.}

\item{resub.err}{Uses the re-substitution error for the calculation of SEs.}

\item{adj.df}{Adjusts the degrees of freedom of the re-substitution error
estimate in the calculation of SEs.}

\item{...}{Additional graphical parameters}
}
\value{
\code{NULL} (invisibly).
}
\description{
Gives a visual representation of the cross-validation results in an
\code{\link{rpartmv-class}} object.
}
\details{
The set of possible cost-complexity prunings of a tree from a nested
set. For the geometric means of the intervals of values of \code{cp} for
which a pruning is optimal, a cross-validation has (usually) been done in
the initial construction by \code{\link{rpartmv}}. Member \code{cptable} of
the \code{\link{rpartmv-class}} contains the mean and standard deviation of
the errors in the cross-validated prediction against each of the geometric
means, and these are plotted by this function. A good choice of \code{cp} for
pruning is often the leftmost value for which the mean lies below the
horizontal line.
}
\seealso{
\code{\link{rpartmv}}, \code{\link{printcp}},
\code{\link{rpartmv-class}}, and \code{\link{mvpart}} (for an example).
}
\keyword{tree}
